+++ node/src/GlideClusterClient.ts
@@ -6,6 +6,8 @@ import * as net from "net";
 import {
     BaseClient,
     BaseClientConfiguration,
+    Decoder,
+    GlideString,
     PubSubMsg,
     ReadFrom, // eslint-disable-line @typescript-eslint/no-unused-vars
     ReturnType,
@@ -341,22 +343,27 @@ export class GlideClusterClient extends BaseClient {
      *  The command will be routed automatically based on the passed command's default request policy, unless `route` is provided,
      *  in which case the client will route the command to the nodes defined by `route`.
      *
-     * See the [Glide for Redis Wiki](https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#custom-command)
+     * Note: An error will occur if the string decoder is used with commands that return only bytes as a response.
+     *
+     * See the [Glide for Valkey Wiki](https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#custom-command)
      * for details on the restrictions and limitations of the custom command API.
      *
      * @example
      * ```typescript
      * // Example usage of customCommand method to retrieve pub/sub clients with routing to all primary nodes
-     * const result = await client.customCommand(["CLIENT", "LIST", "TYPE", "PUBSUB"], "allPrimaries");
+     * const result = await client.customCommand(["CLIENT", "LIST", "TYPE", "PUBSUB"], {route: "allPrimaries", decoder: Decoder.String});
      * console.log(result); // Output: Returns a list of all pub/sub clients
      * ```
      */
     public customCommand(
-        args: string[],
-        route?: Routes,
+        args: GlideString[],
+        options?: { route?: Routes; decoder?: Decoder },
     ): Promise<ClusterResponse<ReturnType>> {
         const command = createCustomCommand(args);
-        return super.createWritePromise(command, toProtobufRoute(route));
+        return super.createWritePromise(command, {
+            route: toProtobufRoute(options?.route),
+            decoder: options?.decoder,
+        });
     }
 
     /** Execute a transaction by processing the queued commands.
@@ -373,11 +380,17 @@ export class GlideClusterClient extends BaseClient {
      */
     public exec(
         transaction: ClusterTransaction,
-        route?: SingleNodeRoute,
+        options?: {
+            route?: SingleNodeRoute;
+            decoder?: Decoder;
+        },
     ): Promise<ReturnType[] | null> {
         return this.createWritePromise<ReturnType[] | null>(
             transaction.commands,
-            toProtobufRoute(route),
+            {
+                route: toProtobufRoute(options?.route),
+                decoder: options?.decoder,
+            },
         ).then((result: ReturnType[] | null) => {
             return this.processResultWithSetCommands(
                 result,
@@ -411,10 +424,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public ping(message?: string, route?: Routes): Promise<string> {
-        return this.createWritePromise(
-            createPing(message),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createPing(message), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /** Get information and statistics about the Redis server.
@@ -433,7 +445,7 @@ export class GlideClusterClient extends BaseClient {
     ): Promise<ClusterResponse<string>> {
         return this.createWritePromise<ClusterResponse<string>>(
             createInfo(options),
-            toProtobufRoute(route),
+            { route: toProtobufRoute(route) },
         );
     }
 
@@ -466,7 +478,7 @@ export class GlideClusterClient extends BaseClient {
     ): Promise<ClusterResponse<string | null>> {
         return this.createWritePromise<ClusterResponse<string | null>>(
             createClientGetName(),
-            toProtobufRoute(route),
+            { route: toProtobufRoute(route) },
         );
     }
 
@@ -486,10 +498,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public configRewrite(route?: Routes): Promise<"OK"> {
-        return this.createWritePromise(
-            createConfigRewrite(),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createConfigRewrite(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /** Resets the statistics reported by Redis using the INFO and LATENCY HISTOGRAM commands.
@@ -508,10 +519,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public configResetStat(route?: Routes): Promise<"OK"> {
-        return this.createWritePromise(
-            createConfigResetStat(),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createConfigResetStat(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /** Returns the current connection id.
@@ -525,7 +535,7 @@ export class GlideClusterClient extends BaseClient {
     public clientId(route?: Routes): Promise<ClusterResponse<number>> {
         return this.createWritePromise<ClusterResponse<number>>(
             createClientId(),
-            toProtobufRoute(route),
+            { route: toProtobufRoute(route) },
         );
     }
 
@@ -560,7 +570,7 @@ export class GlideClusterClient extends BaseClient {
     ): Promise<ClusterResponse<Record<string, string>>> {
         return this.createWritePromise<ClusterResponse<Record<string, string>>>(
             createConfigGet(parameters),
-            toProtobufRoute(route),
+            { route: toProtobufRoute(route) },
         );
     }
 
@@ -585,10 +595,9 @@ export class GlideClusterClient extends BaseClient {
         parameters: Record<string, string>,
         route?: Routes,
     ): Promise<"OK"> {
-        return this.createWritePromise(
-            createConfigSet(parameters),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createConfigSet(parameters), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /** Echoes the provided `message` back.
@@ -617,10 +626,9 @@ export class GlideClusterClient extends BaseClient {
         message: string,
         route?: Routes,
     ): Promise<ClusterResponse<string>> {
-        return this.createWritePromise(
-            createEcho(message),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createEcho(message), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /** Returns the server time.
@@ -650,7 +658,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public time(route?: Routes): Promise<ClusterResponse<[string, string]>> {
-        return this.createWritePromise(createTime(), toProtobufRoute(route));
+        return this.createWritePromise(createTime(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -704,10 +714,10 @@ export class GlideClusterClient extends BaseClient {
         options?: LolwutOptions,
         route?: Routes,
     ): Promise<ClusterResponse<string>> {
-        return this.createWritePromise(
-            createLolwut(options),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createLolwut(options), {
+            decoder: options?.decoder,
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -734,10 +744,9 @@ export class GlideClusterClient extends BaseClient {
         args: string[],
         route?: Routes,
     ): Promise<ReturnType> {
-        return this.createWritePromise(
-            createFCall(func, [], args),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFCall(func, [], args), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -765,10 +774,9 @@ export class GlideClusterClient extends BaseClient {
         args: string[],
         route?: Routes,
     ): Promise<ReturnType> {
-        return this.createWritePromise(
-            createFCallReadOnly(func, [], args),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFCallReadOnly(func, [], args), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -793,10 +801,9 @@ export class GlideClusterClient extends BaseClient {
         libraryCode: string,
         route?: Routes,
     ): Promise<string> {
-        return this.createWritePromise(
-            createFunctionDelete(libraryCode),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFunctionDelete(libraryCode), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -827,7 +834,7 @@ export class GlideClusterClient extends BaseClient {
     ): Promise<string> {
         return this.createWritePromise(
             createFunctionLoad(libraryCode, replace),
-            toProtobufRoute(route),
+            { route: toProtobufRoute(route) },
         );
     }
 
@@ -850,10 +857,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public functionFlush(mode?: FlushMode, route?: Routes): Promise<string> {
-        return this.createWritePromise(
-            createFunctionFlush(mode),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFunctionFlush(mode), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -891,10 +897,9 @@ export class GlideClusterClient extends BaseClient {
         options?: FunctionListOptions,
         route?: Routes,
     ): Promise<ClusterResponse<FunctionListResponse>> {
-        return this.createWritePromise(
-            createFunctionList(options),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFunctionList(options), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -950,10 +955,9 @@ export class GlideClusterClient extends BaseClient {
     public async functionStats(
         route?: Routes,
     ): Promise<ClusterResponse<FunctionStatsResponse>> {
-        return this.createWritePromise(
-            createFunctionStats(),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFunctionStats(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -973,10 +977,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public flushall(mode?: FlushMode, route?: Routes): Promise<string> {
-        return this.createWritePromise(
-            createFlushAll(mode),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFlushAll(mode), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -996,10 +999,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public flushdb(mode?: FlushMode, route?: Routes): Promise<string> {
-        return this.createWritePromise(
-            createFlushDB(mode),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createFlushDB(mode), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -1018,8 +1020,10 @@ export class GlideClusterClient extends BaseClient {
      * console.log("Number of keys across all primary nodes: ", numKeys);
      * ```
      */
-    public dbsize(route?: Routes): Promise<number> {
-        return this.createWritePromise(createDBSize(), toProtobufRoute(route));
+    public dbsize(route?: Routes): Promise<ClusterResponse<number>> {
+        return this.createWritePromise(createDBSize(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /** Publish a message on pubsub channel.
@@ -1214,10 +1218,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async lastsave(route?: Routes): Promise<ClusterResponse<number>> {
-        return this.createWritePromise(
-            createLastSave(),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createLastSave(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -1236,10 +1239,9 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async randomKey(route?: Routes): Promise<string | null> {
-        return this.createWritePromise(
-            createRandomKey(),
-            toProtobufRoute(route),
-        );
+        return this.createWritePromise(createRandomKey(), {
+            route: toProtobufRoute(route),
+        });
     }
 
     /**
@@ -1261,6 +1263,8 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async unwatch(route?: Routes): Promise<"OK"> {
-        return this.createWritePromise(createUnWatch(), toProtobufRoute(route));
+        return this.createWritePromise(createUnWatch(), {
+            route: toProtobufRoute(route),
+        });
     }
 }