+++ node/tests/SharedTests.ts
@@ -22,6 +22,7 @@ import {
     ClosingError,
     ClusterTransaction,
     ConditionalChange,
+    Decoder,
     ExpireOptions,
     FlushMode,
     GeoUnit,
@@ -1251,6 +1252,226 @@ export function runBaseTests<Context>(config: {
                 expect(await client.hget(key, "nonExistingField")).toEqual(
                     null,
                 );
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `hscan test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const key1 = "{key}-1" + uuidv4();
+                const initialCursor = "0";
+                const defaultCount = 20;
+                const resultCursorIndex = 0;
+                const resultCollectionIndex = 1;
+
+                // Setup test data - use a large number of entries to force an iterative cursor.
+                const numberMap: Record<string, string> = {};
+
+                for (let i = 0; i < 50000; i++) {
+                    numberMap[i.toString()] = "num" + i;
+                }
+
+                const charMembers = ["a", "b", "c", "d", "e"];
+                const charMap: Record<string, string> = {};
+
+                for (let i = 0; i < charMembers.length; i++) {
+                    charMap[charMembers[i]] = i.toString();
+                }
+
+                // Result contains the whole set
+                expect(await client.hset(key1, charMap)).toEqual(
+                    charMembers.length,
+                );
+                let result = await client.hscan(key1, initialCursor);
+                expect(result[resultCursorIndex]).toEqual(initialCursor);
+                expect(result[resultCollectionIndex].length).toEqual(
+                    Object.keys(charMap).length * 2, // Length includes the score which is twice the map size
+                );
+
+                const resultArray = result[resultCollectionIndex];
+                const resultKeys = [];
+                const resultValues: string[] = [];
+
+                for (let i = 0; i < resultArray.length; i += 2) {
+                    resultKeys.push(resultArray[i]);
+                    resultValues.push(resultArray[i + 1]);
+                }
+
+                // Verify if all keys from charMap are in resultKeys
+                const allKeysIncluded = resultKeys.every(
+                    (key) => key in charMap,
+                );
+                expect(allKeysIncluded).toEqual(true);
+
+                const allValuesIncluded = Object.values(charMap).every(
+                    (value) => value in resultValues,
+                );
+                expect(allValuesIncluded).toEqual(true);
+
+                // Test hscan with match
+                result = await client.hscan(key1, initialCursor, {
+                    match: "a",
+                });
+
+                expect(result[resultCursorIndex]).toEqual(initialCursor);
+                expect(result[resultCollectionIndex]).toEqual(["a", "0"]);
+
+                // Set up testing data with the numberMap set to be used for the next set test keys and test results.
+                expect(await client.hset(key1, numberMap)).toEqual(
+                    Object.keys(numberMap).length,
+                );
+
+                let resultCursor = initialCursor;
+                const secondResultAllKeys: string[] = [];
+                const secondResultAllValues: string[] = [];
+                let isFirstLoop = true;
+
+                do {
+                    result = await client.hscan(key1, resultCursor);
+                    resultCursor = result[resultCursorIndex].toString();
+                    const resultEntry = result[resultCollectionIndex];
+
+                    for (let i = 0; i < resultEntry.length; i += 2) {
+                        secondResultAllKeys.push(resultEntry[i]);
+                        secondResultAllValues.push(resultEntry[i + 1]);
+                    }
+
+                    if (isFirstLoop) {
+                        expect(resultCursor).not.toBe("0");
+                        isFirstLoop = false;
+                    } else if (resultCursor === initialCursor) {
+                        break;
+                    }
+
+                    // Scan with result cursor has a different set
+                    const secondResult = await client.hscan(key1, resultCursor);
+                    const newResultCursor =
+                        secondResult[resultCursorIndex].toString();
+                    expect(resultCursor).not.toBe(newResultCursor);
+                    resultCursor = newResultCursor;
+                    const secondResultEntry =
+                        secondResult[resultCollectionIndex];
+
+                    expect(result[resultCollectionIndex]).not.toBe(
+                        secondResult[resultCollectionIndex],
+                    );
+
+                    for (let i = 0; i < secondResultEntry.length; i += 2) {
+                        secondResultAllKeys.push(secondResultEntry[i]);
+                        secondResultAllValues.push(secondResultEntry[i + 1]);
+                    }
+                } while (resultCursor != initialCursor); // 0 is returned for the cursor of the last iteration.
+
+                // Verify all data is found in hscan
+                const allSecondResultKeys = Object.keys(numberMap).every(
+                    (key) => key in secondResultAllKeys,
+                );
+                expect(allSecondResultKeys).toEqual(true);
+
+                const allSecondResultValues = Object.keys(numberMap).every(
+                    (value) => value in secondResultAllValues,
+                );
+                expect(allSecondResultValues).toEqual(true);
+
+                // Test match pattern
+                result = await client.hscan(key1, initialCursor, {
+                    match: "*",
+                });
+                expect(result[resultCursorIndex]).not.toEqual(initialCursor);
+                expect(
+                    result[resultCollectionIndex].length,
+                ).toBeGreaterThanOrEqual(defaultCount);
+
+                // Test count
+                result = await client.hscan(key1, initialCursor, {
+                    count: 25,
+                });
+                expect(result[resultCursorIndex]).not.toEqual(initialCursor);
+                expect(
+                    result[resultCollectionIndex].length,
+                ).toBeGreaterThanOrEqual(25);
+
+                // Test count with match returns a non-empty list
+                result = await client.hscan(key1, initialCursor, {
+                    match: "1*",
+                    count: 30,
+                });
+                expect(result[resultCursorIndex]).not.toEqual(initialCursor);
+                expect(result[resultCollectionIndex].length).toBeGreaterThan(0);
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `hscan empty set, negative cursor, negative count, and non-hash key exception tests`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const key1 = "{key}-1" + uuidv4();
+                const key2 = "{key}-2" + uuidv4();
+                const initialCursor = "0";
+                const resultCursorIndex = 0;
+                const resultCollectionIndex = 1;
+
+                // Empty set
+                let result = await client.hscan(key1, initialCursor);
+                expect(result[resultCursorIndex]).toEqual(initialCursor);
+                expect(result[resultCollectionIndex]).toEqual([]);
+
+                // Negative cursor
+                result = await client.hscan(key1, "-1");
+                expect(result[resultCursorIndex]).toEqual(initialCursor);
+                expect(result[resultCollectionIndex]).toEqual([]);
+
+                // Exceptions
+                // Non-hash key
+                expect(await client.set(key2, "test")).toEqual("OK");
+                await expect(client.hscan(key2, initialCursor)).rejects.toThrow(
+                    RequestError,
+                );
+                await expect(
+                    client.hscan(key2, initialCursor, {
+                        match: "test",
+                        count: 20,
+                    }),
+                ).rejects.toThrow(RequestError);
+
+                // Negative count
+                await expect(
+                    client.hscan(key2, initialCursor, {
+                        count: -1,
+                    }),
+                ).rejects.toThrow(RequestError);
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `encoder test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient) => {
+                const key = uuidv4();
+                const value = uuidv4();
+                const valueEncoded = Buffer.from(value);
+
+                expect(await client.set(key, value)).toEqual("OK");
+                expect(await client.get(key)).toEqual(value);
+                expect(await client.get(key, Decoder.Bytes)).toEqual(
+                    valueEncoded,
+                );
+                expect(await client.get(key, Decoder.String)).toEqual(value);
+
+                // Setting the encoded value. Should behave as the previous test since the default is String decoding.
+                expect(await client.set(key, valueEncoded)).toEqual("OK");
+                expect(await client.get(key)).toEqual(value);
+                expect(await client.get(key, Decoder.Bytes)).toEqual(
+                    valueEncoded,
+                );
+                expect(await client.get(key, Decoder.String)).toEqual(value);
             }, protocol);
         },
         config.timeout,
@@ -7886,6 +8107,138 @@ export function runBaseTests<Context>(config: {
                 expect(await client.set(stringKey, "foo")).toEqual("OK");
                 await expect(
                     client.xclaim(stringKey, "_", "_", 0, ["_"]),
+                ).rejects.toThrow(RequestError);
+            }, protocol);
+        },
+        config.timeout,
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        `xautoclaim test_%p`,
+        async (protocol) => {
+            await runTest(async (client: BaseClient, cluster) => {
+                const key = uuidv4();
+                const group = uuidv4();
+
+                expect(
+                    await client.xgroupCreate(key, group, "0", {
+                        mkStream: true,
+                    }),
+                ).toEqual("OK");
+                expect(
+                    await client.xgroupCreateConsumer(key, group, "consumer"),
+                ).toEqual(true);
+
+                expect(
+                    await client.xadd(
+                        key,
+                        [
+                            ["entry1_field1", "entry1_value1"],
+                            ["entry1_field2", "entry1_value2"],
+                        ],
+                        { id: "0-1" },
+                    ),
+                ).toEqual("0-1");
+                expect(
+                    await client.xadd(
+                        key,
+                        [["entry2_field1", "entry2_value1"]],
+                        { id: "0-2" },
+                    ),
+                ).toEqual("0-2");
+
+                expect(
+                    await client.customCommand([
+                        "xreadgroup",
+                        "group",
+                        group,
+                        "consumer",
+                        "STREAMS",
+                        key,
+                        ">",
+                    ]),
+                ).toEqual({
+                    [key]: {
+                        "0-1": [
+                            ["entry1_field1", "entry1_value1"],
+                            ["entry1_field2", "entry1_value2"],
+                        ],
+                        "0-2": [["entry2_field1", "entry2_value1"]],
+                    },
+                });
+
+                let result = await client.xautoclaim(
+                    key,
+                    group,
+                    "consumer",
+                    0,
+                    "0-0",
+                    1,
+                );
+                let expected: typeof result = [
+                    "0-2",
+                    {
+                        "0-1": [
+                            ["entry1_field1", "entry1_value1"],
+                            ["entry1_field2", "entry1_value2"],
+                        ],
+                    },
+                ];
+                if (!cluster.checkIfServerVersionLessThan("7.0.0"))
+                    expected.push([]);
+                expect(result).toEqual(expected);
+
+                let result2 = await client.xautoclaimJustId(
+                    key,
+                    group,
+                    "consumer",
+                    0,
+                    "0-0",
+                );
+                let expected2: typeof result2 = ["0-0", ["0-1", "0-2"]];
+                if (!cluster.checkIfServerVersionLessThan("7.0.0"))
+                    expected2.push([]);
+                expect(result2).toEqual(expected2);
+
+                // add one more entry
+                expect(
+                    await client.xadd(
+                        key,
+                        [["entry3_field1", "entry3_value1"]],
+                        { id: "0-3" },
+                    ),
+                ).toEqual("0-3");
+
+                // incorrect IDs - response is empty
+                result = await client.xautoclaim(
+                    key,
+                    group,
+                    "consumer",
+                    0,
+                    "5-0",
+                );
+                expected = ["0-0", {}];
+                if (!cluster.checkIfServerVersionLessThan("7.0.0"))
+                    expected.push([]);
+                expect(result).toEqual(expected);
+
+                result2 = await client.xautoclaimJustId(
+                    key,
+                    group,
+                    "consumer",
+                    0,
+                    "5-0",
+                );
+                expected2 = ["0-0", []];
+                if (!cluster.checkIfServerVersionLessThan("7.0.0"))
+                    expected2.push([]);
+                expect(result2).toEqual(expected2);
+
+                // key exists, but it is not a stream
+                const stringKey = uuidv4();
+                expect(await client.set(stringKey, "foo")).toEqual("OK");
+                await expect(
+                    client.xautoclaim(stringKey, "_", "_", 0, "_"),
                 ).rejects.toThrow(RequestError);
             }, protocol);
         },