+++ node/tests/TestUtilities.ts
@@ -478,6 +478,76 @@ export function validateTransactionResponse(
 }
 
 /**
+ * Populates a transaction with commands to test the decodable commands with various default decoders.
+ * @param baseTransaction - A transaction.
+ * @param valueEncodedResponse - Represents the encoded response of "value" to compare
+ * @returns Array of tuples, where first element is a test name/description, second - expected return value.
+ */
+export async function encodableTransactionTest(
+    baseTransaction: Transaction | ClusterTransaction,
+    valueEncodedResponse: ReturnType,
+): Promise<[string, ReturnType][]> {
+    const key = "{key}" + uuidv4(); // string
+    const value = "value";
+    // array of tuples - first element is test name/description, second - expected return value
+    const responseData: [string, ReturnType][] = [];
+
+    baseTransaction.set(key, value);
+    responseData.push(["set(key, value)", "OK"]);
+    baseTransaction.get(key);
+    responseData.push(["get(key)", valueEncodedResponse]);
+
+    return responseData;
+}
+
+/**
+ * Populates a transaction with commands to test the decoded response.
+ * @param baseTransaction - A transaction.
+ * @returns Array of tuples, where first element is a test name/description, second - expected return value.
+ */
+export async function encodedTransactionTest(
+    baseTransaction: Transaction | ClusterTransaction,
+): Promise<[string, ReturnType][]> {
+    const key1 = "{key}" + uuidv4(); // string
+    const key2 = "{key}" + uuidv4(); // string
+    const key = "dumpKey";
+    const dumpResult = Buffer.from([
+        0, 5, 118, 97, 108, 117, 101, 11, 0, 232, 41, 124, 75, 60, 53, 114, 231,
+    ]);
+    const value = "value";
+    const valueEncoded = Buffer.from(value);
+    // array of tuples - first element is test name/description, second - expected return value
+    const responseData: [string, ReturnType][] = [];
+
+    baseTransaction.set(key1, value);
+    responseData.push(["set(key1, value)", "OK"]);
+    baseTransaction.set(key2, value);
+    responseData.push(["set(key2, value)", "OK"]);
+    baseTransaction.get(key1);
+    responseData.push(["get(key1)", valueEncoded]);
+    baseTransaction.get(key2);
+    responseData.push(["get(key2)", valueEncoded]);
+
+    baseTransaction.set(key, value);
+    responseData.push(["set(key, value)", "OK"]);
+    baseTransaction.customCommand(["DUMP", key]);
+    responseData.push(['customCommand(["DUMP", key])', dumpResult]);
+    baseTransaction.del([key]);
+    responseData.push(["del(key)", 1]);
+    baseTransaction.get(key);
+    responseData.push(["get(key)", null]);
+    baseTransaction.customCommand(["RESTORE", key, "0", dumpResult]);
+    responseData.push([
+        'customCommand(["RESTORE", key, "0", dumpResult])',
+        "OK",
+    ]);
+    baseTransaction.get(key);
+    responseData.push(["get(key)", valueEncoded]);
+
+    return responseData;
+}
+
+/**
  * Populates a transaction with commands to test.
  * @param baseTransaction - A transaction.
  * @returns Array of tuples, where first element is a test name/description, second - expected return value.
@@ -587,6 +657,13 @@ export async function transactionTest(
     responseData.push(["del([key1])", 1]);
     baseTransaction.hset(key4, { [field]: value });
     responseData.push(["hset(key4, { [field]: value })", 1]);
+    baseTransaction.hscan(key4, "0");
+    responseData.push(['hscan(key4, "0")', ["0", [field, value]]]);
+    baseTransaction.hscan(key4, "0", { match: "*", count: 20 });
+    responseData.push([
+        'hscan(key4, "0", {match: "*", count: 20})',
+        ["0", [field, value]],
+    ]);
     baseTransaction.hstrlen(key4, field);
     responseData.push(["hstrlen(key4, field)", value.length]);
     baseTransaction.hlen(key4);
@@ -965,28 +1042,28 @@ export async function transactionTest(
 
     // key9 has one entry here: {"0-2":[["field","value2"]]}
 
-    baseTransaction.xgroupCreateConsumer(key9, groupName1, "consumer1");
+    baseTransaction.xgroupCreateConsumer(key9, groupName1, consumer);
     responseData.push([
-        "xgroupCreateConsumer(key9, groupName1, consumer1)",
+        "xgroupCreateConsumer(key9, groupName1, consumer)",
         true,
     ]);
     baseTransaction.customCommand([
         "xreadgroup",
         "group",
         groupName1,
-        "consumer1",
+        consumer,
         "STREAMS",
         key9,
         ">",
     ]);
     responseData.push([
-        'xreadgroup(groupName1, "consumer1", key9, >)',
+        "xreadgroup(groupName1, consumer, key9, >)",
         { [key9]: { "0-2": [["field", "value2"]] } },
     ]);
     baseTransaction.xpending(key9, groupName1);
     responseData.push([
         "xpending(key9, groupName1)",
-        [1, "0-2", "0-2", [["consumer1", "1"]]],
+        [1, "0-2", "0-2", [[consumer, "1"]]],
     ]);
     baseTransaction.xpendingWithOptions(key9, groupName1, {
         start: InfScoreBoundary.NegativeInfinity,
@@ -994,44 +1071,55 @@ export async function transactionTest(
         count: 10,
     });
     responseData.push([
-        "xpending(key9, groupName1, -, +, 10)",
-        [["0-2", "consumer1", 0, 1]],
+        "xpendingWithOptions(key9, groupName1, -, +, 10)",
+        [["0-2", consumer, 0, 1]],
     ]);
-    baseTransaction.xclaim(key9, groupName1, "consumer1", 0, ["0-2"]);
+    baseTransaction.xclaim(key9, groupName1, consumer, 0, ["0-2"]);
     responseData.push([
-        'xclaim(key9, groupName1, "consumer1", 0, ["0-2"])',
+        'xclaim(key9, groupName1, consumer, 0, ["0-2"])',
         { "0-2": [["field", "value2"]] },
     ]);
-    baseTransaction.xclaim(key9, groupName1, "consumer1", 0, ["0-2"], {
+    baseTransaction.xclaim(key9, groupName1, consumer, 0, ["0-2"], {
         isForce: true,
         retryCount: 0,
         idle: 0,
     });
     responseData.push([
-        'xclaim(key9, groupName1, "consumer1", 0, ["0-2"], { isForce: true, retryCount: 0, idle: 0})',
+        'xclaim(key9, groupName1, consumer, 0, ["0-2"], { isForce: true, retryCount: 0, idle: 0})',
         { "0-2": [["field", "value2"]] },
     ]);
-    baseTransaction.xclaimJustId(key9, groupName1, "consumer1", 0, ["0-2"]);
+    baseTransaction.xclaimJustId(key9, groupName1, consumer, 0, ["0-2"]);
     responseData.push([
-        'xclaimJustId(key9, groupName1, "consumer1", 0, ["0-2"])',
+        'xclaimJustId(key9, groupName1, consumer, 0, ["0-2"])',
         ["0-2"],
     ]);
-    baseTransaction.xclaimJustId(key9, groupName1, "consumer1", 0, ["0-2"], {
+    baseTransaction.xclaimJustId(key9, groupName1, consumer, 0, ["0-2"], {
         isForce: true,
         retryCount: 0,
         idle: 0,
     });
     responseData.push([
-        'xclaimJustId(key9, groupName1, "consumer1", 0, ["0-2"], { isForce: true, retryCount: 0, idle: 0})',
+        'xclaimJustId(key9, groupName1, consumer, 0, ["0-2"], { isForce: true, retryCount: 0, idle: 0})',
         ["0-2"],
     ]);
-    baseTransaction.xgroupCreateConsumer(key9, groupName1, consumer);
-    responseData.push([
-        "xgroupCreateConsumer(key9, groupName1, consumer)",
-        true,
-    ]);
+
+    if (gte(version, "6.2.0")) {
+        baseTransaction.xautoclaim(key9, groupName1, consumer, 0, "0-0", 1);
+        responseData.push([
+            'xautoclaim(key9, groupName1, consumer, 0, "0-0", 1)',
+            gte(version, "7.0.0")
+                ? ["0-0", { "0-2": [["field", "value2"]] }, []]
+                : ["0-0", { "0-2": [["field", "value2"]] }],
+        ]);
+        baseTransaction.xautoclaimJustId(key9, groupName1, consumer, 0, "0-0");
+        responseData.push([
+            'xautoclaimJustId(key9, groupName1, consumer, 0, "0-0")',
+            gte(version, "7.0.0") ? ["0-0", ["0-2"], []] : ["0-0", ["0-2"]],
+        ]);
+    }
+
     baseTransaction.xgroupDelConsumer(key9, groupName1, consumer);
-    responseData.push(["xgroupDelConsumer(key9, groupName1, consumer)", 0]);
+    responseData.push(["xgroupDelConsumer(key9, groupName1, consumer)", 1]);
     baseTransaction.xgroupDestroy(key9, groupName1);
     responseData.push(["xgroupDestroy(key9, groupName1)", true]);
     baseTransaction.xgroupDestroy(key9, groupName2);
@@ -1070,7 +1158,7 @@ export async function transactionTest(
     baseTransaction.bitpos(key17, 1);
     responseData.push(["bitpos(key17, 1)", 1]);
 
-    if (gte("6.0.0", version)) {
+    if (gte(version, "6.0.0")) {
         baseTransaction.bitfieldReadOnly(key17, [
             new BitFieldGet(new SignedEncoding(5), new BitOffset(3)),
         ]);